Implement Stats & Characters
Created Stat class to track and name different kinds of bounded stats.
Created Effect class to interface with and modify new Stat class.
Created Character class to store character stats, name, and art asset prefabs in a custom asset.
Created a CharacterStore class to manage the central list of characters as a custom asset.
Created a Student class to track dynamic stats, progression, with a constructor that takes the base character.
Created a CharacterSelectRandomiser class to generate sets of random characters for the character select menu.
Created a CharacterSelector class to construct and set the active student from a character.
Created a Character Select Button prefab to use as the base prefab for a button on the character select menu.